+++ db/db_impl/db_impl.h
@@ -1602,7 +1602,7 @@ class DBImpl : public DB {
     // This can be used to limit the synced size tracked in MANIFEST. It is
     // useful for files that undergo append during sync, in which case the file
     // size at tracking time is not necessarily all synced.
-    size_t tracked_synced_size_limit = SIZE_MAX;
+    uint64_t tracked_synced_size_limit = UINT64_MAX;
   };
 
   // PurgeFileInfo is a structure to hold information of files to be deleted in
